requirements.txt
alembic==1.10.4
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.0
fastapi-cache2==0.2.1
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
flower==1.2.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanize==4.6.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.8
packaging==23.1
passlib==1.7.4
pendulum==2.1.2
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
redis==4.5.5
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
